+++ packages/default-layout/__tests__/scrollMode.e2e.ts
@@ -88,4 +88,62 @@ describe('Scroll modes', () => {
         const currentPage = await currentPageInput.evaluate((ele) => ele.getAttribute('value'));
         expect(currentPage).toEqual('7');
     });
+
+    test('Keep the current page after switching the scroll mode', async () => {
+        await page.goto('http://localhost:3000/default-layout');
+        await page.setViewport({
+            width: 1920,
+            height: 1080,
+        });
+        await page.evaluate(() => document.querySelector('[data-testid="core__viewer"]').scrollIntoView());
+
+        // Wait until the first page is rendered
+        await page.waitForSelector('[data-testid="core__text-layer-0"]', { visible: true });
+
+        const pagesContainer = await page.waitForSelector('[data-testid="core__inner-pages"]');
+
+        // Jump to the 4th page
+        let nextPageButton = await page.waitForSelector('[data-testid="page-navigation__next-button"]');
+        await nextPageButton.click();
+        await nextPageButton.click();
+        await nextPageButton.click();
+        await page.waitForSelector('[data-testid="core__text-layer-3"]', { visible: true });
+
+        let scrollTop = await pagesContainer.evaluate((ele) => ele.scrollTop);
+        expect(scrollTop).toEqual(3612);
+
+        // Switch to the horizontal scroll mode
+        let moreActionsButton = await page.waitForSelector('[data-testid="toolbar__more-actions-popover-target"]');
+        await moreActionsButton.click();
+        const switchToHorizontalMenu = await page.waitForSelector('[data-testid="scroll-mode__horizontal-menu"]');
+        await switchToHorizontalMenu.click();
+
+        let scrollLeft = await pagesContainer.evaluate((ele) => ele.scrollLeft);
+        expect(scrollLeft).toEqual(2721);
+
+        // Zoom to 50%
+        const zoomButton = await page.waitForSelector('[data-testid="zoom__popover-target"]');
+        await zoomButton.click();
+
+        const zoomPopover = await page.waitForSelector('[id="rpv-core__popover-body-inner-zoom"]');
+        const zoomMenuItem = await zoomPopover.waitForSelector('button:nth-of-type(4)');
+        const zoomLevel = await zoomMenuItem.evaluate((ele) => ele.textContent);
+        expect(zoomLevel).toEqual('50%');
+        await zoomMenuItem.click();
+
+        scrollLeft = await pagesContainer.evaluate((ele) => ele.scrollLeft);
+        expect(scrollLeft).toEqual(907);
+
+        // Switch to the wrapped scroll mode
+        moreActionsButton = await page.waitForSelector('[data-testid="toolbar__more-actions-popover-target"]');
+        await moreActionsButton.click();
+        const switchToHWrappedMenu = await page.waitForSelector('[data-testid="scroll-mode__wrapped-menu"]');
+        await switchToHWrappedMenu.click();
+
+        scrollTop = await pagesContainer.evaluate((ele) => ele.scrollTop);
+        expect(scrollTop).toEqual(412);
+
+        scrollLeft = await pagesContainer.evaluate((ele) => ele.scrollLeft);
+        expect(scrollLeft).toEqual(0);
+    });
 });